refactor(parser): split stmt.rs and expr.rs into submodules
stmt.rs (3,588 lines) → stmt/{mod,class,enum_decl,trait_use}.rs
expr.rs (3,025 lines) → expr/{mod,atom}.rs
interpolation.rs → expr/interpolation.rs

No behavior changes. All external crate paths preserved via re-exports.

Closes #274
